Update core_mapping interface configuration
If interfaces are added to the CVMX container outside of docker
control (e.g. by stitching veth-pairs to interconnect CVMX instances),
the interfaces were not detected during startup and as a consequence
RIOT did not startup correctly.

Added new configuration script that is called after all interfaces are
attached and up.